[ia32][x64] Make SharedTurboAssembler a class template
Previously SharedTurboAssembler was a base class for ia32 and x64
TurboAssembler. This made it easy to share code, only if the
implementation was the same. In some cases, like
ExternalReferenceAsOperand, the implementation defers slightly between
the two architectures. By converting SharedTurboAssembler into a class
template, and following the CRTP, SharedTurboAssembler has access to the
actual class via the template parameter, and can call
architecture-specific macro-assembler functions.

Bug: v8:11589
Change-Id: I9319bd0c26c76995cef43ae5ec3f69392b3f825b