heap: ArrayBufferSweeper: Fix freed counter overflow
The array buffer sweeper has its own freed counter which is migrated
back to a global counter. There exist two paths for finalizing array
buffer sweeping which both need to merge back the counters. If we miss
out on merging back the counter, the freed counter may overflow in the
next cycle.

Bug: chromium:1241332
Change-Id: Ic985f72414198de2eaf900b8e2e9b39bed24d87e