[x64][ia32] Share Pinsrb/Pinsrw/Pinsrd code
Pinsrb, Pinsrw was separately defined in ia32 and x64 macro assembler,
move them into shared-macro-assembler.

Pinsrd is slightly tricky,
since it is also used in JS, and the pre-SSE4.1 instructions are
different across backend. We share the AVX and SSE4_1 code, but delegate
to base class for the pre-SSE4.1 code (via CRTP).

Pinsrq is only defined on x64, but requires the PinsrHelper, so we move
PinsrHelper into shared-macro-assembler header.

Bug: v8:11589
Change-Id: Ib5f7a626cbbdcd74d773df86de12de79a9e5fd76