[sandbox] Introduce CagedPointer
A CagedPointer is guaranteed to point into the Virtual Memory Cage and
is used for ArrayBuffer backing stores when the heap sandbox is enabled.
In the current implementation, CagedPointers are stored as offsets from
the cage base, shifted to the left. Because the cage base address is
usually available in a register, accessing a CagedPointer is very
efficient, requiring only an additional shift and add operation.

CagedPointers cannot generally represent nullptr, as NULL usually lies
outside the cage. As such, nullptr backing stores are replaced with a
special empty backing store value, which, in the current implementation,
points to the end of the cage, right in front of the trailing guard
regions. Due to this, it is no longer correct to compare a backing store
pointer against nullptr.

Bug: chromium:1218005
Change-Id: Ifc8c088e3862400672051a8c52840514dee2911f
Cq-Include-Trybots: luci.v8.try:v8_linux64_heap_sandbox_dbg_ng