[sandbox] Introduce CagedPointer
A CagedPointer is guaranteed to point into the Virtual Memory Cage and
is used mainly for ArrayBuffer backing stores. In the current
implementation, this property is achieved by storing CagedPointers as
offsets from the cage base, shifted to the left. Because the cage base
address is usually available in a register, "uncaging" a CagedPointer is
a very efficient operation.

Since a CagedPointer cannot generally represent nullptr, special
handling for nullptr values becomes necessary. However, this is not
required on performance critical paths as the pointers must be
non-nullptr when accessed there.

Bug: chromium:1218005
Change-Id: Ifc8c088e3862400672051a8c52840514dee2911f
Cq-Include-Trybots: luci.v8.try:v8_linux64_heap_sandbox_dbg_ng