[compiler] Fix inconsistency between JSFunctionRef and FeedbackCellRef
It could happen that the information about the feedback vector cached in
a JSFunctionData disagreed with the current value of the function's
feedback cell. The inlining code wasn't prepared for that and a CHECK
could fail.

The CL fixes this, mainly by removing the caching of
has_feedback_vector and feedback_vector.

Bug: v8:12172
Change-Id: Ife3ab8872085d9496e6d1f34514114a086f653ad