[builtins] Unify the logic in ArrayPrototypeSlice and ReduceArrayPrototypeSlice
They need to agree about when to delegate to CloneFastJSArray, since it
produces arrays which are potentially COW. If they don't agree, TF
generates code which produces a COW array and then expects it to be
non-COW -> immediate deopt.

This CL gets rid of the discrepancy in the case when there's exactly
one argument and it's the number 0.

Some corner cases remain:
- 1st argument not a number but ToInteger returns 0
- 2nd argument exists but its value is 'undefined'

These should be extremely rare in the real world.

Bug:v8:12194
Change-Id: I10230245c97f8997da4d79702f29ebff11297229