[turbofan] AllocateRaw cannot observe StoreField
StoreStoreElimination has a whitelist of IrOpcodes that can't observe
StoreFields. Before this CL this did not include kAllocateRaw. That
meant that e.g., repeated double stores to the same field could not be
eliminated if they required the double to be boxed (an allocation). By
adding it we can e.g., eliminate repeated double stores to the same
context field.

Bug: v8:12200
Change-Id: Ifb4cfa554e20fa3b92f5aefa1a7c7899f77f3171