[heap] Allocate less space for markbits
Currently, each MemoryChunk header contains enough space to include
markbits for every location within the chunk, including locations that
are part of the header. We don't need the ability to mark whether the
header data is reachable, so we can avoid allocating some of that space,
which leaves more of the chunk available for allocated object data. This
change reduces the MemoryChunk header size by 248 bytes on builds with
pointer compression enabled. 248 bytes out of each 256k chunk amounts to
saving roughly 0.1% of the total heap memory (or less in large-object
chunks). This is clearly not a huge improvement, but it's an easy way to
shrink the heap by a small amount.

Change-Id: Iacc472e6c92f1537b3e388f8eee7377d0afb0f2c