[x64] Optimize v128.const when two int64 halves are the same
Previously it generates a mvoq+pinsrq, now it generates a single
movddup.

movddup is smaller in code size, and also faster on most arch (latency
1, rthroughput of 1, 1 uop, uses 1 port) than pinsrq (latency 2, 2 uop,
uses 2 port) (from https://uops.info/table.html)

Moddup is mean to work on FP domain, so we might incur 1 cycle
forwarding penalty, but for v128.const, it could be used as in input to
in or float. So overall, such mixed-domain issues should be neutral with
the prior codegen (pinsrq).

Bug: v8:11033
Change-Id: Iab81168ffad84488b90ff307d440bed15c9f90a3