[snapshots] Add a new snapshot for shareable objects
This CL adds a new snapshot to hold shareable objects, i.e. objects that
are in the shared heap or may need to be in the shared heap depending on
runtime flags. Currently this is to support --shared-string-table, which
puts all in-place-internalizable strings, internalized strings, and the
string table into the shared heap.

The shareable snapshot is never deserialized into client Isolates. This
means when V8 is started without a shared Isolate, the shareable
snapshot is deserialized into all Isolates.

Bug: v8:12007
Change-Id: I7eeab73080cda2e8250a5a49747f25b2440a349d