heap: Fix for V8 histograms and persistent memory
Until now, histograms associated with isolate counters were created
at the time of isolate initialization. This is too early because it
happens before persistent memory is configured. Histograms created
before persistent memory is set up are retrieved periodically but
infrequently and are not "flushed" on process termination. As a result,
a lot of samples from V8 are lost.

This CL delays the creation of counter histograms until the context
has been initialized.

Bug: chromium:1270428
Change-Id: I10e46fc7f0666777e74ac51875d7da962a6b7ab6