[parser] Allow cloning chunked streams
This allows streamed sources to also trigger parallel compile tasks. The
chunk vectors themselves are copied, and their data is shared via
std::shared_ptr.

Cloned chunked streams are initialised with a null source, and are not
allowed to fetch any more data. This is ok for the purposes of cloning
for parallel compile tasks, as the relevant function will have already
been preparsed and thus its data will already have been fetched.

Change-Id: I58f24faf01b06c00901f8f4d3377d8e3a407c051