cppgc: CppHeap no longer inherits from EmbedderHeapTracer api
This CL does the following:
1) Split out of EmbedderHeapTracer a base class EmbedderHeapTracerBase
that CppHeap inherits from. This class holds shared functionality to
avoid reimplementing it in CppHeap.
2) Remove registration of CppHeap as a remote tracer.
3) Revise LocalEmbedderHeapTracer as a switching point between CppHeap
and a remote tracer. Currently it is assumed that CppHeap and a remote
tracer are mutually exclusive and only one can be used at any given
time.
4) Mark v8::Isolate::GetEmbedderHeapTracer as soon to deprecated since
it can no longer account for CppHeap.

Bug: v8:12407
Change-Id: I53513d181ab63f56a88f05c3b76b47ac4dffe86f