[cleanup][codegen] Fix -Wshadow warnings (attempt 2)
The fix in https://crrev.com/c/3283620 doesn't work, it was meant to be
a general way to fix shadow whenever ASM_CODE_COMMENT is used, by
appending the current line, via __LINE__ macro to the variable name.
However, __LINE__ is not expended correctly when it is directly part of
a variable name.

To fix the shadowing, we instead give a different name for the function
level code comment that uses __func__ as the comment. This is sufficient
to cover all the warnings we have now (and should be sufficient for the
future, it is unlikely we want 2 function comments, since they will
be identical).

Bug: v8:12244,v8:12245
Change-Id: I7b9b9da74d62023ac631ff0876f41e80971e2f09